+++ usecode/ucdisasm.cc
@@ -44,7 +44,7 @@ using std::cout;
 int Usecode_internal::get_opcode_length(int opcode)
 {
 	if (opcode >=0 &&
-		opcode < (sizeof(opcode_table)/sizeof(opcode_table[0]))) {
+		(unsigned)opcode < (sizeof(opcode_table)/sizeof(opcode_table[0]))) {
 
 	    return opcode_table[opcode].nbytes + 1;
 	} else {
@@ -73,7 +73,7 @@ void Usecode_internal::uc_trace_disasm(Usecode_value* locals, int num_locals,
 	uint8* param_ip = ip;
 	_opcode_desc* pdesc;
 
-	if (opcode >=0 && opcode < (sizeof(opcode_table)/sizeof(opcode_table[0])))
+	if (opcode >=0 && (unsigned)opcode < (sizeof(opcode_table)/sizeof(opcode_table[0])))
 		pdesc = &(opcode_table[opcode]);
 	signed short immed;
 	uint16 varref;
@@ -111,7 +111,7 @@ void Usecode_internal::uc_trace_disasm(Usecode_value* locals, int num_locals,
 			case DATA_STRING32:
 				{
 					char* pstr;
-					int len;
+					size_t len;
 					// Print data string operand
 					if (pdesc->type == DATA_STRING)
 						offset = Read2(ip);
@@ -123,7 +123,7 @@ void Usecode_internal::uc_trace_disasm(Usecode_value* locals, int num_locals,
 					if( len > 20 )
 						len = 20 - 3;
 					std::printf("\tL%04X\t\t; ", offset);
-					for(int i = 0; i < len; i++ )
+					for(unsigned i = 0; i < len; i++ )
 						std::printf("%c", pstr[i]);
 					if( len < strlen(pstr) )
 						// String truncated
@@ -214,12 +214,12 @@ void Usecode_internal::uc_trace_disasm(Usecode_value* locals, int num_locals,
 				std::printf("[%04X]\t= ", (uint16)staticref);
 				if (staticref < 0) {
 					// global
-					if (-staticref < statics.size())
+					if ((unsigned)(-staticref) < statics.size())
 						statics[-staticref].print(cout, true);
 					else
 						std::printf("<out of range>");
 				} else {
-					if (staticref < locstatics.size())
+					if ((unsigned)staticref < locstatics.size())
 						locstatics[staticref].print(cout, true);
 					else
 						std::printf("<out of range>");
